feat [pepeny] adding sql schema and make utils to store function
