+++ commands/kas.js
@@ -1,18 +1,5 @@
-const monthsMap = {
-    1: 'Januari', 2: 'Februari', 3: 'Maret', 4: 'April',
-    5: 'Mei', 6: 'Juni', 7: 'Juli', 8: 'Agustus',
-    9: 'September', 10: 'Oktober', 11: 'November', 12: 'Desember'
-};
-
-function formattedKas(num) {
-    if (num >= 1000) {
-        let formatted = (num / 1000).toFixed(1);
-        formatted = formatted.replace(/\.0$/, '');
-        return `${formatted}K`;
-    }
-    return num.toString();
-}
-
+const monthsMap = require('../utils/monthsMap');
+const formattedMoney = require('../utils/formattedMoney');
 module.exports = {
     name: 'kas',
     description: 'All about kas command',
@@ -23,11 +10,11 @@ module.exports = {
     ],
     async execute(message, args, db, client) {
         if (!args[0]) {
-            return message.reply('Tolong sebutin subcommand atau nama, contoh: !kas total / !kas history Andi / !kas {nama}');
+            return message.reply('!kas bisa jalan kalo lu sebutin subcommand atau nama, contoh: !kas total / !kas history {nama} / !kas {nama}');
         }
 
         const sub = args[0].toLowerCase();
-
+        // !kas total 
         if (sub === 'total') {
             db.query('SELECT SUM(jumlah) AS total FROM kas', (err, results) => {
                 if (err) return message.reply('Error: ' + err.message);
@@ -36,11 +23,12 @@ module.exports = {
                 const embed = client.baseEmbed(message)
                     .setColor(0x00AE86)
                     .setTitle('💰 Kas Kelas')
-                    .setDescription(`Total kas saat ini: **${formattedKas(total)}**`);
+                    .setDescription(`Total kas saat ini: **${formattedMoney(total)}**`);
 
                 message.reply({ embeds: [embed] });
             });
         }
+        // !kas history {nama}
         else if (sub === 'history') {
             const nama = args.slice(1).join(' ');
             if (!nama) return message.reply('Kurang sebutin nama, contoh: !kas history Andi');
@@ -54,7 +42,8 @@ module.exports = {
                     const bulanText = monthsMap[r.bulan] || r.bulan;
                     const tgl = new Date(r.tanggal);
                     const formattedDate = `${String(tgl.getDate()).padStart(2,'0')} ${bulanText} ${tgl.getFullYear()} ${String(tgl.getHours()).padStart(2,'0')}:${String(tgl.getMinutes()).padStart(2,'0')}`;
-                    historyText += `Bulan ${bulanText} Minggu ${r.minggu}: ${formattedKas(r.jumlah)} dibayar tgl ${formattedDate}\n`;
+                    historyText += r.jumlah == 0 ? `Bulan ${bulanText} Minggu ${r.minggu}: Belum dibayar ❌\n` 
+                                                : `Bulan ${bulanText} Minggu ${r.minggu}: ${formattedMoney(r.jumlah)} (${formattedDate})\n`;
                 });
 
                 const embed = client.baseEmbed(message)
@@ -65,11 +54,13 @@ module.exports = {
                 message.reply({ embeds: [embed] });
             });
         }
+        // !kas {nama}
         else {
             const nama = args.join(' ');
-            db.query('SELECT * FROM kas WHERE nama = ? ORDER BY bulan, minggu', [nama], (err, results) => {
+            const currMonth = new Date().getMonth() + 1;
+            db.query('SELECT * FROM kas WHERE nama = ? AND bulan = ? ORDER BY minggu', [nama, currMonth], (err, results) => {
                 if (err) return message.reply('Error: ' + err.message);
-                if (results.length === 0) return message.reply(`Data kas si **${nama}** gaada.`);
+                if (results.length === 0) return message.reply(`Data kas si **${nama}** di bulan ini gaada.`);
 
                 const absen = results[0].absen;
                 const bulan = results[0].bulan;
@@ -89,12 +80,12 @@ module.exports = {
                 }
 
                 let finalStatus = notYetPaid > 0
-                    ? `\nTotal kas yang belum dibayar : **${formattedKas(notYetPaid)}**\n`
-                    : `\nKas ${nama} udah lunas!\n`;
+                    ? `\nTotal kas yang belum dibayar : **${formattedMoney(notYetPaid)}**\n`
+                    : `\nKas ${nama} bulan ini udah lunas!\n`;
 
                 const embed = client.baseEmbed(message)
                     .setColor(0x3498db)
-                    .setTitle(`📊 Status Kas ${nama}`)
+                    .setTitle(`📊 Status Kas ${nama} di bulan ${bulanToText}`)
                     .addFields(
                         { name: 'Nama', value: nama, inline: true },
                         { name: 'Absen', value: absen.toString(), inline: true },
