+++ database/seed.sql
@@ -1,0 +1,12 @@
+INSERT INTO kas (nama, absen, jumlah, bulan, minggu)
+VALUES
+('Andi', 1, 5000, 9, 1),
+('Andi', 1, 5000, 9, 2),
+('Andi', 1, 5000, 9, 3),
+('Budi', 2, 5000, 9, 1),
+('Budi', 2, 0,    9, 2),
+('Citra', 3, 5000, 9, 1),
+('Citra', 3, 5000, 9, 2),
+('Dewi', 4, 0,    9, 1),
+('Eko', 5, 5000, 9, 1),
+('Eko', 5, 5000, 9, 2);
